# tsdb.PostingsForMatchers: Optimize inverse matching