--- a/tsdb/block.go
+++ b/tsdb/block.go
@@ -83,7 +83,7 @@ type IndexReader interface {
 	PostingsForLabelMatching(ctx context.Context, name string, match func(value string) bool) index.Postings
 
 	// PostingsForAllLabelValues returns a sorted iterator over all postings having a label with the given name.
-	// If no postings are found having at least one matching label, an empty iterator is returned.
+	// If no postings are found with the label in question, an empty iterator is returned.
 	PostingsForAllLabelValues(ctx context.Context, name string) index.Postings
 
 	// SortedPostings returns a postings list that is reordered to be sorted