Simple follow-up to #13620. Just modify `tsdb.PostingsForMatchers` to use the optimized `tsdb.IndexReader.PostingsForLabelMatching` method also for inverse matching.

The performance is much improved for a subset of the cases; there are up to ~60% CPU gains and ~12.5% reduction in memory usage. Please consult benchmark statistics below.

Removing `TestReader_InversePostingsForMatcherHonorsContextCancel` since `inversePostingsForMatcher` only passes `ctx` to `IndexReader` implementations now.

#### Benchmark results

<details>
<summary>PostingsForMatchers benchmark stats</summary>

```console
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Intel(R) Xeon(R) Platinum 8358 CPU @ 2.60GHz
                                                                     │ head-postings-main.txt │ head-postings-inverse-postings-for-matchers.txt │
                                                                     │         sec/op         │          sec/op           vs base               │
Querier/Head/PostingsForMatchers/n="1"-8                                         458.0n ±  1%               454.5n ±  3%        ~ (p=0.119 n=6)
Querier/Head/PostingsForMatchers/n="X"-8                                         455.2n ±  0%               449.6n ±  0%   -1.23% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",j="foo"-8                                 723.6n ±  0%               711.5n ±  0%   -1.67% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="X",j="foo"-8                                 511.0n ±  0%               511.2n ±  0%        ~ (p=0.584 n=6)
Querier/Head/PostingsForMatchers/j="foo",n="1"-8                                 723.8n ±  0%               712.4n ±  0%   -1.57% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",j!="foo"-8                                702.8n ±  0%               690.0n ±  0%   -1.83% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!="2"-8                                  717.8n ±  0%               702.6n ±  0%   -2.11% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="X",j!="foo"-8                                489.4n ±  0%               488.2n ±  0%        ~ (p=0.132 n=6)
Querier/Head/PostingsForMatchers/i=~"1[0-9]",j=~"foo|bar"-8                      1.409µ ±  1%               1.411µ ±  1%        ~ (p=0.604 n=6)
Querier/Head/PostingsForMatchers/j=~"foo|bar"-8                                  802.8n ±  0%               796.1n ±  0%   -0.83% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/j=~"XXX|YYY"-8                                  655.1n ±  0%               658.4n ±  0%   +0.50% (p=0.009 n=6)
Querier/Head/PostingsForMatchers/j=~"X.+"-8                                      605.6n ±  0%               604.3n ±  0%   -0.20% (p=0.041 n=6)
Querier/Head/PostingsForMatchers/i=~"(1|2|3|4|5|6|20|55)"-8                      1.144µ ±  1%               1.119µ ±  0%   -2.23% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i!~"(1|2|3|4|5|6|20|55)"-8                      1.266µ ±  2%               1.245µ ±  1%   -1.62% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~"X|Y|Z"-8                                    740.2n ±  1%               739.0n ±  0%        ~ (p=0.240 n=6)
Querier/Head/PostingsForMatchers/i!~"X|Y|Z"-8                                    848.2n ±  1%               845.5n ±  1%        ~ (p=0.223 n=6)
Querier/Head/PostingsForMatchers/i=~".*"-8                                       416.0n ±  0%               413.1n ±  0%   -0.71% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~"1.*"-8                                      9.378m ±  6%               7.934m ±  6%  -15.39% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~".*1"-8                                      6.716m ±  1%               6.187m ±  9%   -7.88% (p=0.015 n=6)
Querier/Head/PostingsForMatchers/i=~".+"-8                                       27.34m ±  4%               26.46m ±  6%        ~ (p=0.093 n=6)
Querier/Head/PostingsForMatchers/i=~".+",j=~"X.+"-8                              27.22m ±  3%               25.73m ±  6%   -5.46% (p=0.026 n=6)
Querier/Head/PostingsForMatchers/i=~""-8                                         28.30m ±  4%               11.67m ± 15%  -58.75% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i!=""-8                                         28.47m ±  2%               11.48m ±  2%  -59.68% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".*",j="foo"-8                         778.2n ±  0%               764.6n ±  1%   -1.75% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="X",i=~".*",j="foo"-8                         557.9n ±  0%               559.0n ±  0%        ~ (p=0.310 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".*",i!="2",j="foo"-8                  1.034µ ±  0%               1.031µ ±  0%        ~ (p=0.229 n=6)
Querier/Head/PostingsForMatchers/n="1",i!=""-8                                   28.03m ±  2%               11.36m ±  8%  -59.48% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!="",j="foo"-8                           28.34m ±  3%               11.64m ±  2%  -58.91% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!="",j=~"X.+"-8                          28.57m ±  2%               11.39m ±  4%  -60.11% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!="",j=~"XXX|YYY"-8                      28.45m ±  1%               11.19m ±  3%  -60.68% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~"X|Y|Z",j="foo"-8                      1.033µ ±  0%               1.041µ ±  0%   +0.82% (p=0.004 n=6)
Querier/Head/PostingsForMatchers/n="1",i!~"X|Y|Z",j="foo"-8                      1.218µ ±  0%               1.221µ ±  0%        ~ (p=0.128 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",j="foo"-8                         27.66m ±  4%               26.18m ±  5%   -5.33% (p=0.009 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~"1.+",j="foo"-8                        9.340m ±  3%               7.358m ± 14%  -21.22% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".*1.*",j="foo"-8                      23.75m ±  2%               21.91m ±  6%   -7.73% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!="2",j="foo"-8                  27.64m ±  2%               25.85m ±  4%   -6.47% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~"2.*",j="foo"-8                36.84m ±  6%               34.57m ±  6%   -6.15% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~".*2.*",j="foo"-8              50.36m ±  3%               48.14m ±  6%   -4.41% (p=0.041 n=6)
Querier/Head/PostingsForMatchers/n="X",i=~".+",i!~".*2.*",j="foo"-8              648.4n ±  0%               653.9n ±  0%   +0.85% (p=0.002 n=6)
Querier/Block/PostingsForMatchers/n="1"-8                                        35.12µ ±  1%               34.69µ ±  0%   -1.22% (p=0.002 n=6)
Querier/Block/PostingsForMatchers/n="X"-8                                        595.4n ±  4%               626.1n ±  2%   +5.14% (p=0.009 n=6)
Querier/Block/PostingsForMatchers/n="1",j="foo"-8                                387.9µ ±  2%               384.8µ ±  2%        ~ (p=0.699 n=6)
Querier/Block/PostingsForMatchers/n="X",j="foo"-8                                607.2n ±  1%               623.6n ±  1%   +2.69% (p=0.002 n=6)
Querier/Block/PostingsForMatchers/j="foo",n="1"-8                                384.8µ ±  1%               384.8µ ±  2%        ~ (p=1.000 n=6)
Querier/Block/PostingsForMatchers/n="1",j!="foo"-8                               385.0µ ±  1%               385.9µ ±  2%        ~ (p=0.699 n=6)
Querier/Block/PostingsForMatchers/n="1",i!="2"-8                                 35.57µ ±  0%               35.63µ ±  0%   +0.16% (p=0.017 n=6)
Querier/Block/PostingsForMatchers/n="X",j!="foo"-8                               568.8n ±  2%               591.8n ±  2%   +4.05% (p=0.002 n=6)
Querier/Block/PostingsForMatchers/i=~"1[0-9]",j=~"foo|bar"-8                     6.005µ ±  0%               5.963µ ±  1%        ~ (p=0.058 n=6)
Querier/Block/PostingsForMatchers/j=~"foo|bar"-8                                 966.3µ ±  3%               951.5µ ±  5%        ~ (p=0.699 n=6)
Querier/Block/PostingsForMatchers/j=~"XXX|YYY"-8                                 552.1n ±  2%               570.9n ±  0%   +3.40% (p=0.002 n=6)
Querier/Block/PostingsForMatchers/j=~"X.+"-8                                     578.8n ±  0%               584.9n ±  0%   +1.05% (p=0.002 n=6)
Querier/Block/PostingsForMatchers/i=~"(1|2|3|4|5|6|20|55)"-8                     4.220µ ±  0%               4.186µ ±  1%   -0.79% (p=0.002 n=6)
Querier/Block/PostingsForMatchers/i!~"(1|2|3|4|5|6|20|55)"-8                     956.9µ ± 17%               961.0µ ±  4%        ~ (p=0.818 n=6)
Querier/Block/PostingsForMatchers/i=~"X|Y|Z"-8                                   686.4n ±  2%               693.8n ±  2%        ~ (p=0.418 n=6)
Querier/Block/PostingsForMatchers/i!~"X|Y|Z"-8                                   979.8µ ± 10%               927.2µ ± 25%        ~ (p=0.180 n=6)
Querier/Block/PostingsForMatchers/i=~".*"-8                                      980.2µ ±  7%               959.1µ ±  4%        ~ (p=0.093 n=6)
Querier/Block/PostingsForMatchers/i=~"1.*"-8                                     3.798m ±  1%               3.795m ±  1%        ~ (p=0.937 n=6)
Querier/Block/PostingsForMatchers/i=~".*1"-8                                     2.514m ±  1%               2.537m ±  1%   +0.91% (p=0.009 n=6)
Querier/Block/PostingsForMatchers/i=~".+"-8                                      17.66m ±  2%               17.69m ±  4%        ~ (p=0.589 n=6)
Querier/Block/PostingsForMatchers/i=~".+",j=~"X.+"-8                             17.67m ±  2%               17.87m ±  1%        ~ (p=0.132 n=6)
Querier/Block/PostingsForMatchers/i=~""-8                                        20.89m ±  2%               16.61m ±  1%  -20.46% (p=0.002 n=6)
Querier/Block/PostingsForMatchers/i!=""-8                                        19.13m ±  2%               14.81m ±  1%  -22.59% (p=0.002 n=6)
Querier/Block/PostingsForMatchers/n="1",i=~".*",j="foo"-8                        384.9µ ±  1%               386.1µ ±  2%        ~ (p=0.485 n=6)
Querier/Block/PostingsForMatchers/n="X",i=~".*",j="foo"-8                        640.7n ±  1%               661.5n ±  1%   +3.24% (p=0.002 n=6)
Querier/Block/PostingsForMatchers/n="1",i=~".*",i!="2",j="foo"-8                 386.7µ ±  1%               386.2µ ±  1%        ~ (p=0.485 n=6)
Querier/Block/PostingsForMatchers/n="1",i!=""-8                                  19.14m ±  2%               14.92m ±  1%  -22.04% (p=0.002 n=6)
Querier/Block/PostingsForMatchers/n="1",i!="",j="foo"-8                          19.97m ±  0%               15.67m ±  1%  -21.53% (p=0.002 n=6)
Querier/Block/PostingsForMatchers/n="1",i!="",j=~"X.+"-8                         19.22m ±  1%               14.81m ±  2%  -22.95% (p=0.002 n=6)
Querier/Block/PostingsForMatchers/n="1",i!="",j=~"XXX|YYY"-8                     19.30m ±  1%               14.95m ±  2%  -22.51% (p=0.002 n=6)
Querier/Block/PostingsForMatchers/n="1",i=~"X|Y|Z",j="foo"-8                     35.13µ ±  0%               35.18µ ±  0%        ~ (p=0.132 n=6)
Querier/Block/PostingsForMatchers/n="1",i!~"X|Y|Z",j="foo"-8                     385.5µ ±  3%               385.9µ ±  1%        ~ (p=0.699 n=6)
Querier/Block/PostingsForMatchers/n="1",i=~".+",j="foo"-8                        18.47m ±  2%               18.71m ±  2%        ~ (p=0.310 n=6)
Querier/Block/PostingsForMatchers/n="1",i=~"1.+",j="foo"-8                       4.555m ±  2%               4.601m ±  3%        ~ (p=0.699 n=6)
Querier/Block/PostingsForMatchers/n="1",i=~".*1.*",j="foo"-8                     11.20m ±  1%               11.11m ±  1%        ~ (p=0.093 n=6)
Querier/Block/PostingsForMatchers/n="1",i=~".+",i!="2",j="foo"-8                 18.54m ±  1%               18.70m ±  1%        ~ (p=0.093 n=6)
Querier/Block/PostingsForMatchers/n="1",i=~".+",i!~"2.*",j="foo"-8               22.38m ±  2%               22.81m ±  2%   +1.95% (p=0.004 n=6)
Querier/Block/PostingsForMatchers/n="1",i=~".+",i!~".*2.*",j="foo"-8             29.12m ±  2%               29.55m ±  1%   +1.47% (p=0.041 n=6)
Querier/Block/PostingsForMatchers/n="X",i=~".+",i!~".*2.*",j="foo"-8             863.0n ±  1%               862.3n ±  3%        ~ (p=0.855 n=6)
geomean                                                                          130.5µ                     118.1µ         -9.52%

                                                                     │ head-postings-main.txt │ head-postings-inverse-postings-for-matchers.txt │
                                                                     │          B/op          │          B/op           vs base                 │
Querier/Head/PostingsForMatchers/n="1"-8                                           64.00 ± 0%               64.00 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="X"-8                                           48.00 ± 0%               48.00 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",j="foo"-8                                   176.0 ± 0%               176.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="X",j="foo"-8                                   48.00 ± 0%               48.00 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/j="foo",n="1"-8                                   176.0 ± 0%               176.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",j!="foo"-8                                  176.0 ± 0%               176.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i!="2"-8                                    176.0 ± 0%               176.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="X",j!="foo"-8                                  48.00 ± 0%               48.00 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~"1[0-9]",j=~"foo|bar"-8                        512.0 ± 0%               512.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/j=~"foo|bar"-8                                    352.0 ± 0%               352.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/j=~"XXX|YYY"-8                                    128.0 ± 0%               128.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/j=~"X.+"-8                                        80.00 ± 0%               80.00 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~"(1|2|3|4|5|6|20|55)"-8                        416.0 ± 0%               416.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i!~"(1|2|3|4|5|6|20|55)"-8                        480.0 ± 0%               480.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~"X|Y|Z"-8                                      176.0 ± 0%               176.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i!~"X|Y|Z"-8                                      240.0 ± 0%               240.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~".*"-8                                         64.00 ± 0%               64.00 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~"1.*"-8                                      2.722Mi ± 0%             2.722Mi ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~".*1"-8                                      1.531Mi ± 0%             1.531Mi ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~".+"-8                                       12.22Mi ± 0%             12.22Mi ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~".+",j=~"X.+"-8                              12.22Mi ± 0%             12.22Mi ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/i=~""-8                                         12.22Mi ± 0%             10.69Mi ± 0%  -12.53% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i!=""-8                                         12.22Mi ± 0%             10.69Mi ± 0%  -12.53% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".*",j="foo"-8                           176.0 ± 0%               176.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="X",i=~".*",j="foo"-8                           48.00 ± 0%               48.00 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".*",i!="2",j="foo"-8                    288.0 ± 0%               288.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i!=""-8                                   12.22Mi ± 0%             10.69Mi ± 0%  -12.53% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!="",j="foo"-8                           12.22Mi ± 0%             10.69Mi ± 0%  -12.53% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!="",j=~"X.+"-8                          12.22Mi ± 0%             10.69Mi ± 0%  -12.53% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i!="",j=~"XXX|YYY"-8                      12.22Mi ± 0%             10.69Mi ± 0%  -12.53% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~"X|Y|Z",j="foo"-8                        240.0 ± 0%               240.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i!~"X|Y|Z",j="foo"-8                        352.0 ± 0%               352.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".+",j="foo"-8                         12.22Mi ± 0%             12.22Mi ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~"1.+",j="foo"-8                        2.722Mi ± 0%             2.722Mi ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".*1.*",j="foo"-8                      5.906Mi ± 0%             5.906Mi ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!="2",j="foo"-8                  12.22Mi ± 0%             12.22Mi ± 0%        ~ (p=1.000 n=6) ¹
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~"2.*",j="foo"-8                14.95Mi ± 0%             14.95Mi ± 0%        ~ (p=1.000 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~".*2.*",j="foo"-8              18.14Mi ± 0%             18.14Mi ± 0%        ~ (p=1.000 n=6)
Querier/Head/PostingsForMatchers/n="X",i=~".+",i!~".*2.*",j="foo"-8                48.00 ± 0%               48.00 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Block/PostingsForMatchers/n="1"-8                                          80.00 ± 0%               80.00 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Block/PostingsForMatchers/n="X"-8                                          64.00 ± 0%               64.00 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Block/PostingsForMatchers/n="1",j="foo"-8                                  208.0 ± 0%               208.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Block/PostingsForMatchers/n="X",j="foo"-8                                  64.00 ± 0%               64.00 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Block/PostingsForMatchers/j="foo",n="1"-8                                  208.0 ± 0%               208.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Block/PostingsForMatchers/n="1",j!="foo"-8                                 208.0 ± 0%               208.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Block/PostingsForMatchers/n="1",i!="2"-8                                   208.0 ± 0%               208.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Block/PostingsForMatchers/n="X",j!="foo"-8                                 64.00 ± 0%               64.00 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Block/PostingsForMatchers/i=~"1[0-9]",j=~"foo|bar"-8                       352.0 ± 0%               352.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Block/PostingsForMatchers/j=~"foo|bar"-8                                   352.0 ± 0%               352.0 ± 0%        ~ (p=1.000 n=6) ¹
Querier/Block/PostingsForMatchers/j=~"XXX|YYY"-8                                   96.00 ± 0%               96.00 ± 0%        ~ (p=1.000 n=6) ¹
…