@aknuds1 2024-09-15
Moving out of draft mode, since AFAICT the cases @colega was unsure about are correct (they are for matchers matching the empty string).