@aknuds1 2024-09-18
Done benchmarking. After rebasing on main, performance regressions disappeared (due to recent regexp optimizations in the PromQL engine I think).