@aknuds1 2024-09-18
> I didn't follow why the test is removed. It didn't work, or the condition is covered by a different test?

@bboreham the reason for removing`TestReader_InversePostingsForMatcherHonorsContextCancel` is that after my changes, the tested function, `inversePostingsForMatcher`, no longer itself honours context cancellation. All it does is pass `ctx` on to `IndexReader` methods. Therefore, the test has lost its raison d'être. Does that make sense?