@aknuds1 2024-11-18
> but I would consider adding a new explicit method instead of changing the behaviour of the existing one, to stay on the safe side.

Thanks @colega, done!